Wrote draft design hypothesis + questions for playtesters
Notes about the making of the prototype:

What is it about?

Goal: Build paths and bridges to collect the coins
How: Creating platforms to reach floating coins
Visual: Starry Night with glowing lights/snow

Visible instruction: Use the platform to create
Hidden instruction: Collect coins → counter to help 

A question that will be answered through playtesting and that you do not know the answer to in advance! The prototype should be built for this purpose (the question that the prototype is built to answer for yourself) (Edit: a yes/no question that is not "can i do this" or "is it possible to do this". In other words: the question that will be answered by the playtest)